Split up 3mf implementation into separate files for v1, v2 and disabled
